Implement deleting teacher and reviewer

+++ ReviewVisualizer.WebApi/Controllers/TeacherController.cs
@@ -76,9 +76,14 @@ namespace ReviewVisualizer.WebApi.Controllers
             {
                 return NotFound();
             }
-            
+
             lock (_lock)
             {
+                if (CheckIfTeacherIsUnderReview(teacher))
+                {
+                    return BadRequest("Cannot delete teacher. It is under active review");
+                }
+
                 Review[] reviews = _dbContext.Reviews.Where(r => r.TeacherId == teacher.Id).ToArray();
                 _dbContext.Reviews.RemoveRange(reviews);
 
@@ -201,5 +206,15 @@ namespace ReviewVisualizer.WebApi.Controllers
                 return Ok(grade);
             }
         }
+
+        #region == Helper methods ==
+
+        private bool CheckIfTeacherIsUnderReview(Teacher teacher)
+        {
+            var reviewers = _dbContext.Reviewers.Where(r => r.Teachers.Any(t => t.Id == teacher.Id));
+
+            return reviewers?.Any(r => !r.IsStopped) ?? false;
+        }
+        #endregion
     }
 }
